Corrected typedef for u32

+++ sensor_api_common_types.h
@@ -184,7 +184,7 @@
 	/*unsigned integer types*/
 	typedef unsigned char u8;/**< used for unsigned 8bit */
 	typedef unsigned short int u16;/**< used for unsigned 16bit */
-	typedef unsigned long int u32;/**< used for unsigned 32bit */
+	typedef unsigned int u32;/**< used for unsigned 32bit */
 
 	/* If your machine support 32 bit
 	define the MACHINE_32_BIT*/
@@ -198,7 +198,7 @@
 	/*unsigned integer types*/
 	typedef unsigned char u8;/**< used for unsigned 8bit */
 	typedef unsigned short int u16;/**< used for unsigned 16bit */
-	typedef unsigned long int u32;/**< used for unsigned 32bit */
+	typedef unsigned int u32;/**< used for unsigned 32bit */
 	typedef unsigned long long int u64;/**< used for unsigned 64bit */
 
 	/* If your machine support 64 bit
@@ -258,7 +258,7 @@
 	/*unsigned integer types*/
 	typedef unsigned char u8;/**< used for unsigned 8bit */
 	typedef unsigned short int u16;/**< used for unsigned 16bit */
-	typedef unsigned long int u32;/**< used for unsigned 32bit */
+	typedef unsigned int u32;/**< used for unsigned 32bit */
 	/*! @brief If your machine support 32 bit
 	define the MACHINE_32_BIT*/
 #elif defined MACHINE_32_BIT
@@ -271,7 +271,7 @@
 	/*unsigned integer types*/
 	typedef unsigned char u8;/**< used for unsigned 8bit */
 	typedef unsigned short int u16;/**< used for unsigned 16bit */
-	typedef unsigned long  int u32;/**< used for unsigned 32bit
+	typedef unsigned int u32;/**< used for unsigned 32bit
 	- int and long int is same for u32*/
 	typedef unsigned long long int u64;/**< used for unsigned 64bit */
